requirements.txt
discord.py==1.6.0
asyncio
requests==2.24.0
python-dotenv
